+++ R/AllClasses.R
@@ -1,11 +1,17 @@
 # CLASSES DEFINITION AND INITIALIZATION
-#' @include reexport.R
+
+# Import classes ===============================================================
+#' @importClassesFrom aion RataDie
+#' @importClassesFrom aion TimeSeries
+#' @importClassesFrom aion TimeIntervals
 NULL
 
 # MCMC =========================================================================
 #' MCMC
 #'
 #' An S4 class to represent the output of a MCMC algorithm.
+#' @slot .Data A [`numeric`] `matrix` giving the MCMC samples expressed in
+#'  *[rata die][aion::RataDie-class]*.
 #' @slot labels A [`character`] vector specifying the name of the events.
 #' @slot depth A [`numeric`] vector giving the sample depth.
 #' @slot hash A [`character`] string giving the 32-byte MD5 hash of the
@@ -37,7 +43,7 @@ NULL
 ## Events ----------------------------------------------------------------------
 #' MCMC Events
 #'
-#' S4 classes to represent a collection of events.
+#' An S4 class to represent a collection of events.
 #' @note
 #'  This class inherits from [`MCMC-class`].
 #' @author N. Frerebeau
@@ -63,7 +69,7 @@ NULL
 ## Phases ----------------------------------------------------------------------
 #' MCMC Phases
 #'
-#' S4 classes to represent a collection of phases.
+#' An S4 class to represent a collection of phases.
 #' @slot labels A [`character`] vector specifying the name of the phases.
 #' @slot hash A [`character`] string giving the 32-byte MD5 hash of the
 #'  original data file.
@@ -71,7 +77,7 @@ NULL
 #'  A phase object is an\eqn{n x m x 2}{n \times m \times 2} array, with
 #'  \eqn{n} being the number of iterations, \eqn{m} being the number of phases
 #'  and with the 2 columns of the third dimension containing the boundaries of
-#'  the phases.
+#'  the phases expressed in *[rata die][aion::RataDie-class]*.
 #' @section Subset:
 #'  In the code snippets below, `x` is a `PhasesMCMC` object.
 #'  \describe{
@@ -96,19 +102,24 @@ NULL
 )
 
 # Time Range ===================================================================
-#' Cumulative Events
+#' Time Range
 #'
 #' An S4 class to represent time ranges.
-#' @slot start,end A `numeric` [`matrix`] giving the lower and upper
-#'  boundaries.
-#' @slot labels A [`character`] vector specifying the name of the events/phases.
 #' @slot hash A [`character`] string giving the 32-byte MD5 hash of the
 #'  original data file.
 #' @section Coerce:
-#'  In the code snippets below, `x` is a `CumulativeEvents` object.
+#'  In the code snippets below, `x` is a `TimeRange` object.
 #'  \describe{
 #'   \item{`as.data.frame(x)`}{Coerces to a [`data.frame`].}
 #'  }
+#' @section Plot:
+#'  In the code snippets below, `x` is a `TimeRange` object.
+#'  \describe{
+#'   \item{`plot(x)`}{Results in a graphic being displayed
+#'   (invisibly returns `x`).}
+#'  }
+#' @note
+#'  This class inherits from [`aion::TimeIntervals-class`].
 #' @author N. Frerebeau
 #' @family classes
 #' @docType class
@@ -117,21 +128,19 @@ NULL
 .TimeRange <- setClass(
   Class = "TimeRange",
   slots = c(
-    start = "matrix",
-    end = "matrix",
-    labels = "matrix",
     hash = "character"
-  )
+  ),
+  contains = "TimeIntervals"
 )
 
 # Tempo ========================================================================
 #' Cumulative Events
 #'
-#' An S4 class to store the result of a [`tempo`] plot.
+#' An S4 class to store the result of a [tempo][tempo()] plot.
 #' @slot lower A [`numeric`] vector giving the lower boundaries of the
-#'  credibility interval.
+#'  credibility interval expressed in *[rata die][aion::RataDie-class]*.
 #' @slot upper A [`numeric`] vector giving the upper boundaries of the
-#'  credibility interval.
+#'  credibility interval expressed in *[rata die][aion::RataDie-class]*.
 #' @slot level A length-one [`numeric`] vector giving the confidence level.
 #' @slot gauss A [`logical`] scalar indicating if the Gaussian approximation of
 #'  the credible interval was used.
@@ -146,7 +155,7 @@ NULL
 #'   \item{`as.data.frame(x)`}{Coerces to a [`data.frame`].}
 #'  }
 #' @note
-#'  This class inherits from [`TimeSeries-class`].
+#'  This class inherits from [`aion::TimeSeries-class`].
 #' @author N. Frerebeau
 #' @family classes
 #' @docType class
@@ -168,7 +177,7 @@ NULL
 # Activity =====================================================================
 #' Activity
 #'
-#' An S4 class to store the result of an [`activity`] plot.
+#' An S4 class to store the result of an [activity][activity()] plot.
 #' @slot hash A [`character`] string giving the 32-byte MD5 hash of the
 #'  original data file.
 #' @section Coerce:
@@ -177,7 +186,7 @@ NULL
 #'   \item{`as.data.frame(x)`}{Coerces to a [`data.frame`].}
 #'  }
 #' @note
-#'  This class inherits from [`TimeSeries-class`].
+#'  This class inherits from [`aion::TimeSeries-class`].
 #' @author N. Frerebeau
 #' @family classes
 #' @docType class
@@ -194,12 +203,8 @@ NULL
 # Occurrence ===================================================================
 #' Occurrence
 #'
-#' An S4 class to store the result of an [`occurrence`] plot.
+#' An S4 class to store the result of an [occurrence][occurrence()] plot.
 #' @slot events An [`integer`] vector giving the occurrence.
-#' @slot start A [`numeric`] vector giving the lower boundaries of the
-#'  credibility interval.
-#' @slot end A [`numeric`] vector giving the upper boundaries of the
-#'  credibility interval.
 #' @slot level A length-one [`numeric`] vector giving the confidence level.
 #' @slot hash A [`character`] string giving the 32-byte MD5 hash of the
 #'  original data file.
@@ -208,6 +213,14 @@ NULL
 #'  \describe{
 #'   \item{`as.data.frame(x)`}{Coerces to a [`data.frame`].}
 #'  }
+#' @section Plot:
+#'  In the code snippets below, `x` is a `OccurrenceEvents` object.
+#'  \describe{
+#'   \item{`plot(x)`}{Results in a graphic being displayed
+#'   (invisibly returns `x`).}
+#'  }
+#' @note
+#'  This class inherits from [`aion::TimeIntervals-class`].
 #' @author N. Frerebeau
 #' @family classes
 #' @docType class
@@ -217,11 +230,10 @@ NULL
   Class = "OccurrenceEvents",
   slots = c(
     events = "integer",
-    start = "numeric",
-    end = "numeric",
     level = "numeric",
     hash = "character"
-  )
+  ),
+  contains = "TimeIntervals"
 )
 
 # Age-Depth Model ==============================================================