+++ R/ArchaeoPhases-internal.R
@@ -1,7 +1,8 @@
 # HELPERS
 
-`%||%` <- function(x, y) {
-  if (!is.null(x)) x else y
+## https://michaelchirico.github.io/potools/articles/developers.html
+tr_ <- function(...) {
+  enc2utf8(gettext(paste0(...), domain = "R-aion"))
 }
 
 # periode <- Vectorize(