+++ R/coerce.R
@@ -5,7 +5,7 @@ NULL
 # To data.frame ================================================================
 #' @export
 #' @method as.data.frame CumulativeEvents
-as.data.frame.CumulativeEvents <- function(x, ..., calendar = getOption("ArchaeoPhases.calendar")) {
+as.data.frame.CumulativeEvents <- function(x, ..., calendar = get_calendar()) {
   tmp <- data.frame(
     time = aion::time(x, calendar = calendar),
     estimate = x[, 1, 1, drop = TRUE]
@@ -34,7 +34,7 @@ setMethod("as.data.frame", "CumulativeEvents", as.data.frame.CumulativeEvents)
 
 #' @export
 #' @method as.data.frame ActivityEvents
-as.data.frame.ActivityEvents <- function(x, ..., calendar = getOption("ArchaeoPhases.calendar")) {
+as.data.frame.ActivityEvents <- function(x, ..., calendar = get_calendar()) {
   methods::callNextMethod() # Method for 'TimeSeries'
 }
 
@@ -45,11 +45,11 @@ setMethod("as.data.frame", "ActivityEvents", as.data.frame.ActivityEvents)
 
 #' @export
 #' @method as.data.frame OccurrenceEvents
-as.data.frame.OccurrenceEvents <- function(x, ..., calendar = getOption("ArchaeoPhases.calendar")) {
+as.data.frame.OccurrenceEvents <- function(x, ..., calendar = get_calendar()) {
   data.frame(
     events = x@events,
-    start = aion::as_year(x@start, calendar = calendar),
-    end = aion::as_year(x@end, calendar = calendar)
+    start = aion::start(x, calendar = calendar),
+    end = aion::end(x, calendar = calendar)
   )
 }
 
@@ -60,22 +60,14 @@ setMethod("as.data.frame", "OccurrenceEvents", as.data.frame.OccurrenceEvents)
 
 #' @export
 #' @method as.data.frame TimeRange
-as.data.frame.TimeRange <- function(x, ..., calendar = getOption("ArchaeoPhases.calendar")) {
-
-  ok <- !is.na(x@start)
-  start <- x@start[ok]
-  end <- x@end[ok]
-  duration <- abs(end - start)
-
-  ## Change calendar
-  if (!is.null(calendar) && length(ok) > 0) {
-    start <- aion::as_year(start, calendar = calendar)
-    end <- aion::as_year(end, calendar = calendar)
-    duration <- aion::as_year(duration, calendar = calendar)
-  }
-
-  data.frame(start = start, end = end, duration = duration,
-             row.names = x@labels[ok])
+as.data.frame.TimeRange <- function(x, ..., calendar = get_calendar()) {
+  ## Build a data frame
+  data.frame(
+    label = labels(x),
+    start = aion::start(x, calendar = calendar),
+    end = aion::end(x, calendar = calendar),
+    duration = aion::span(x, calendar = calendar)
+  )
 }
 
 #' @export
@@ -92,7 +84,7 @@ setMethod(
   signature = "MCMC",
   definition = function(from, chains = 1) {
     ## Validation
-    arkhe::needs("coda")
+    arkhe::assert_package("coda")
 
     L <- nrow(from) / chains
     obj <- vector(mode = "list", length = chains)