+++ inst/tinytest/test_activity.R
@@ -1,3 +1,5 @@
+Sys.setenv(LANGUAGE = "en") # Force locale
+
 # Activity plot ================================================================
 eve <- as_events(mcmc_events, calendar = CE(), iteration = 1)
 eve <- eve[1:1000, ]
@@ -9,11 +11,7 @@ expect_error(current = activity(tempo(eve, count = TRUE)),
 
 if (at_home()) {
   using("tinysnapshot")
-  options(tinysnapshot_device = "svglite")
-  options(tinysnapshot_height = 7) # inches
-  options(tinysnapshot_width = 7)
-  options(tinysnapshot_tol = 300) # pixels
-  options(tinysnapshot_os = "Linux")
+  source("helpers.R")
 
   plot_activity <- function() plot(act)
   expect_snapshot_plot(plot_activity, "plot_activity")