+++ inst/tinytest/test_events.R
@@ -1,3 +1,5 @@
+Sys.setenv(LANGUAGE = "en") # Force locale
+
 eve <- as_events(mcmc_events, calendar = CE(), iteration = 1)
 
 names(eve) <- c("A", "B", "C", "D")
@@ -5,11 +7,7 @@ expect_identical(names(eve), colnames(eve))
 
 if (at_home()) {
   using("tinysnapshot")
-  options(tinysnapshot_device = "svglite")
-  options(tinysnapshot_height = 7) # inches
-  options(tinysnapshot_width = 7)
-  options(tinysnapshot_tol = 1000) # pixels
-  options(tinysnapshot_os = "Linux")
+  source("helpers.R")
 
   ## Events --------------------------------------------------------------------
   plot_event <- function() plot(eve, interval = NULL, sort = FALSE)